move ProfileDropDown files to new folder
