+++ components/ProfileDropdownMenu/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './ProfileDropdownMenu'
