+++ components/ProfileDropdownMenu/profileDropDown.module.scss
@@ -1,6 +1,6 @@
 @use 'sass:color';
-@use '_variables.module.scss';
-@import 'colors.module.scss';
+@use '../../scss/variables.module';
+@import '../../scss/colors.module.scss';
 
 .nav-user-toggle {
   display: flex;
